+++ business web/components/sub/SkillDataProvider.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import React from 'react'
-import { motion } from 'framer-motion'
+import {motion} from 'framer-motion'
 import { useInView } from 'react-intersection-observer'
 import Image from 'next/image';
 
@@ -12,34 +12,34 @@ interface Props {
     index: number;
 }
 
-const SkillDataProvider = ({ src, width, height, index }: Props) => {
-    const { ref, inView } = useInView({
+const SkillDataProvider = ({ src, width, height, index} : Props) => {
+    const {ref, inView} = useInView({
         triggerOnce: true
     })
 
     const imageVariants = {
-        hidden: { opacity: 0 },
-        visible: { opacity: 1 }
+        hidden: {opacity: 0},
+        visible: {opacity: 1}
     }
 
     const animationDelay = 0.3
-    return (
-        <motion.div
-            ref={ref}
-            initial="hidden"
-            variants={imageVariants}
-            animate={inView ? "visible" : "hidden"}
-            custom={index}
-            transition={{ delay: index * animationDelay }}
-        >
-            <Image
-                src={src}
-                width={width}
-                height={height}
-                alt='skill image'
-            />
-        </motion.div>
-    )
+  return (
+  <motion.div
+  ref={ref}
+  initial="hidden"
+  variants={imageVariants}
+  animate={inView ? "visible" : "hidden"}
+  custom={index}
+  transition={{delay: index * animationDelay}}
+  >
+    <Image
+src={src}
+width={width}
+height={height}
+alt='skill image'
+    />
+  </motion.div>
+  )
 }
 
 export default SkillDataProvider
