+++ business web/tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "components/sub/ProjectCard.js", "components/sub/projectdetailtab.js", "components/sub/teamDetails.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "components/sub/ProjectCard.js", "components/sub/projectdetailtab.js"],
   "exclude": ["node_modules"]
 }
